+++ src/main/java/xreliquary/entities/EntityEnderShot.java
@@ -66,7 +66,7 @@ public class EntityEnderShot extends Entity implements IProjectile {
 
 	@Override
 	protected void entityInit() {
-		dataWatcher.addObject(16, Byte.valueOf((byte) 0));
+		dataWatcher.addObject(16, (byte) 0);
 	}
 
 	/**
